+++ tests/other_test.go
@@ -3,10 +3,7 @@ package test
 
 import (
 	"github.com/stretchr/testify/assert"
-	"github.com/terraform-ibm-modules/ibmcloud-terratest-wrapper/common"
 	"github.com/terraform-ibm-modules/ibmcloud-terratest-wrapper/testhelper"
-	"log"
-	"os"
 	"testing"
 )
 
@@ -19,38 +16,6 @@ func TestRunRestoredDBExample(t *testing.T) {
 		Testing:      t,
 		TerraformDir: restoredTerraformDir,
 		Prefix:       "pg-backup",
-	})
-
-	output, err := options.RunTestConsistency()
-	assert.Nil(t, err, "This should not have errored")
-	assert.NotNil(t, output, "Expected some output")
-}
-
-var permanentResources map[string]interface{}
-
-// TestMain will be run before any parallel tests, used to read data from yaml for use with tests
-func TestMain(m *testing.M) {
-
-	var err error
-	permanentResources, err = common.LoadMapFromYaml(yamlLocation)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	os.Exit(m.Run())
-}
-
-func TestRunPointInTimeRecoveryDBExample(t *testing.T) {
-	t.Parallel()
-
-	options := testhelper.TestOptionsDefaultWithVars(&testhelper.TestOptions{
-		Testing:      t,
-		TerraformDir: pitrTerraformDir,
-		Prefix:       "pg-pitr",
-		TerraformVars: map[string]interface{}{
-			"pitr_id": permanentResources["postgresqlCrn"],
-			"region":  region,
-		},
 	})
 
 	output, err := options.RunTestConsistency()
